ai/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
google-cloud-documentai>=2.21
google-cloud-aiplatform>=1.38
google-cloud-vision
google-cloud-translate<3.0.0
google-cloud-texttospeech
google-cloud-storage
faiss-cpu>=1.7
numpy>=1.24
scikit-learn>=1.2
python-dotenv>=1.0
pydantic>=2.4
typing-extensions>=4.8
python-docx>=1.0.1
requests>=2.31
python-docx>=1.0.1
